Fix: Configuration파트 변경
1. Parser함수별 파일 생성
2. SimpleException클래스 구조 변경
3. 테스터기 수정 및 케이스 추가

Related to #45
