+++ config_tester/Simple/AllowMethod/duplicate.conf
