+++ config_tester/Simple/AllowMethod/invalid.conf
