+++ config_tester/Simple/AllowMethod/ok.conf
