+++ config_tester/Simple/Listen/StrangeChar1.conf
