+++ config_tester/Simple/Listen/StrangeChar2.conf
