+++ config_tester/Simple/TryFiles/uri.conf
