+++ config_tester/Simple/Unknown/NoHost.conf
@@ -1,0 +1,4 @@
+server
+{
+	laisten 127.0.0.1:80;
+}
