--- config_tester/SimpleTest/Listen/NoArgument.conf
@@ -1,4 +1,0 @@
-server
-{
-	listen ;
-}
