--- config_tester/SimpleTest/Listen/NoHost.conf
@@ -1,4 +1,0 @@
-server
-{
-	listen :80;
-}
