--- config_tester/SimpleTest/Listen/OnlyHost.conf
@@ -1,4 +1,0 @@
-server
-{
-	listen 127.0.0.1;
-}
