--- config_tester/SimpleTest/Listen/OnlyPort.conf
@@ -1,4 +1,0 @@
-server
-{
-	listen 80;
-}
