--- config_tester/SimpleTest/Listen/StrangeChar1.conf
@@ -1,4 +1,0 @@
-server
-{
-	listen 127a.0.0.1;
-}
