--- config_tester/SimpleTest/Listen/StrangeChar2.conf
@@ -1,4 +1,0 @@
-server
-{
-	listen 127a;
-}
