--- config_tester/SimpleTest/Listen/listen.conf
@@ -1,4 +1,0 @@
-server
-{
-	listen 127.0.0.1:80;
-}
