+++ config_tester/Syntax/Brace/onlyBrace.conf
